requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
kombu==5.3.7
multidict==6.0.5
packaging==24.0
phonenumbers==8.13.35
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
twilio==9.0.5
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
